address keith feedback
